cleanup

+++ course/src/utils.rs
@@ -1,9 +1,9 @@
 use crate::vec3::Vec3;
 
-pub fn create_grid(x_size: usize, y_size: usize) -> Vec<(usize, usize)> {
+pub fn create_grid(x_size: u32, y_size: u32) -> Vec<(u32, u32)> {
     (0..x_size)
-        .flat_map(|x| (0..y_size).map(move |y| (x as usize, y as usize)))
-        .collect::<Vec<(usize, usize)>>()
+        .flat_map(|x| (0..y_size).map(move |y| (x, y)))
+        .collect::<Vec<(u32, u32)>>()
 }
 
 pub fn to0_255_color_format(color: Vec3) -> [u8; 3] {
